[lexical] Perf: Build out the bench infrastructure for first-class use
- Move shared bench helpers to `_utils.ts` (microbench) and `dom/_utils.ts`
  (real-editor) so each bench file can stay focused on the workload.
- Switch the project file selection from `test.include` to
  `test.benchmark.include`. In bench mode vitest reads
  `benchmark.include`; the previous config let vitest fall back to its
  default test glob and pick up files outside the project's intended
  directory.
- Expand the README to cover: why we have two projects (node vs jsdom),
  filtering with `--project`, the comparison-pattern conventions, DCE
  prevention, and how to read the summary block.
